+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - xx.xx.xxxx
+### Fixed
+- Testthat calls with namespace not being parsed
+
 ## [0.5.1] - 21.02.2022
 ### Fixed
 - R reporter missing from the .vsix package